extract into separate test module
