+++ src/moonlink/src/storage/compaction.rs
@@ -3,3 +3,5 @@ pub(super) mod table_compaction;
 
 #[cfg(test)]
 mod test_utils;
+#[cfg(test)]
+mod tests;
